MAINT: Move internal writers to submodule
