+++ openff/system/interop/internal/lammps.py
@@ -1,0 +1,16 @@
+from pathlib import Path
+from typing import Union
+
+from openff.system.components.system import System
+
+
+def to_lammps(openff_sys: System, file_path: Union[Path, str]):
+    assert openff_sys is not None
+
+    if isinstance(file_path, str):
+        path = Path(file_path)
+    if isinstance(file_path, Path):
+        path = file_path
+
+    with open(path, "w") as top_file:
+        top_file.write("; Generated by OpenFF System\n")
